Spring course. Data JPA finished

+++ SchoolApplication/src/main/java/ua/com/foxminded/javaspring/SchoolApplication/command/commands/RemoveStudentByIdCommand.java
@@ -4,12 +4,15 @@ import java.util.List;
 import java.util.Optional;
 import java.util.Scanner;
 
+import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 
+import jakarta.transaction.Transactional;
 import ua.com.foxminded.javaspring.SchoolApplication.db.service.StudentService;
 import ua.com.foxminded.javaspring.SchoolApplication.model.Student;
 
 @Service
+@Component
 public class RemoveStudentByIdCommand implements Command {
 
 	public static final String COMMAND_NAME = "remove_student_by_id";
@@ -18,11 +21,11 @@ public class RemoveStudentByIdCommand implements Command {
 	private final Scanner scanner = new Scanner(System.in);
 
 	public RemoveStudentByIdCommand(StudentService studentService) {
-
 		this.studentService = studentService;
 	}
 
 	@Override
+	@Transactional
 	public void execute() {
 
 		List<Student> students = studentService.findAll();
@@ -32,10 +35,20 @@ public class RemoveStudentByIdCommand implements Command {
 
 		int id = scanner.nextInt();
 
-		Optional<Student> student = studentService.findById(id);
-		studentService.delete(student.get());
+		scanner.nextLine();
 
-		System.out.println(student.get().getName() + ", Id: " + student.get().getKey() + "-  was removed");
+		Optional<Student> student = studentService.findById(id);
+
+		if (student.isPresent()) {
+
+			studentService.delete(student.get());
+			System.out.println(student.get().getName() + ", Id: " + student.get().getKey() + "-  was removed");
+
+		} else {
+			System.out.println("Student is absent");
+
+		}
+		scanner.close();
 	}
 
 	@Override
